+++ src/com/intellij/idea/plugin/hybris/common/HybrisConstants.kt
@@ -86,7 +86,6 @@ object HybrisConstants {
     const val EXTENSION_INFO_XML = "extensioninfo.xml"
     const val EXTENSIONS_XML = "extensions.xml"
     const val COCKPIT_NG_CONFIG_XML = "-config.xml"
-    const val COCKPIT_NG_WIDGETS_XML = "widgets.xml"
     const val COCKPIT_NG_DEFINITION_XML = "definition.xml"
 
     const val HYBRIS_DIRECTORY = "hybris"
@@ -323,6 +322,8 @@ object HybrisConstants {
     const val COCKPIT_NG_NAMESPACE_KEY = "COCKPIT_NG_NAMESPACE"
     const val COCKPIT_NG_INITIALIZE_CONTEXT_TYPE = "ctx.TYPE_CODE"
     const val COCKPIT_NG_WIDGET_ID_STUB = "STUB_"
+
+    const val SCHEMA_COCKPIT_NG_WIDGETS = "http://www.hybris.com/schema/cockpitng/widgets.xsd"
 
     val DEFAULT_JUNK_FILE_NAMES = listOf(
         ".classpath",