+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## [2023.2.10]
 
+### `Cockpit NG` enhancements
+- Improved detection of the `widgets` files [#759](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/759)
+
 ### Other
 - Upgraded Gradle `8.2.1` -> `8.4` [#758](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/758)
 - Make the Plugin incompatible with other plugins [#757](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/757)